tests/wpt/mozilla/tests/mozilla/createEvent-storageevent.html

<!doctype html>
<meta charset="utf-8">
<title>storageevent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_equals(document.createEvent('storageevent').url, "");
}, "document.createEvent('storageevent') does not propagate the document's URL.")
</script>
